python_worker/requirements.txt
python-dotenv
pymysql
pandas
openpyxl